Constrained decoding in RNNG based on JSON schema
Summary:
Inside an intent, restrict the set of actions to the valid slots according to the specified ontology.

This is an attempt to refactor out the ontology reading part in D8043688. The `OntologyConstraint` reads an ontology from the JSON file specified in the `constraints.ontology` config.

For a flow run, `constraints.ontology` can also be one of "dev", "staging", or "prod", in which case the appropriate ontology JSON will be copied from the configerator.

Reviewed By: sonalgupta

Differential Revision: D9127553

fbshipit-source-id: 0dd1f0613603b123fbe70246154ad0cb7045aaee