Register jobspec explictly
Summary: currently we use jobspec decorator to register jobspec, however the file that define the actual jobspec still has to be imported in order to execute the registering code. The import statement must be in particular order, which is unclear and error prone. This diff change the registering to a explicit way.

Reviewed By: bethebunny

Differential Revision: D9306107

fbshipit-source-id: f9e8e4827bcda56ec41bfd1297c45984b74b9620