Add depth based frame accuracies in PyText metrics
Summary:
Add function compute_frame_accuracies_by_depth() which computes bucketized frame accuracies according to the depth of the gold annotation tree. See task T32873635. These metrics are to be added to workflow UI in D9566114 (among other things).

The type FramePredictionPair defined in previous diffs (which represents a pair of predicted/expected frames) is not used where it should be in many places. This causes compute_frame_accuracies_by_depth() to fail. We use it whenever possible in this diff.

pytext/metrics/metrics.py is getting bloated, we will have a diff to split it into multiple files.

Reviewed By: sonalgupta

Differential Revision: D9599298

fbshipit-source-id: 40c3554b575ed6b2291afe4c485badab1f1d026d