Featurizer Interfaces: 1/X, No Serialized Thrift Inputs
Summary:
This diff introduces the base class `Featurizer`, along with new input structs. For the moment, the return type is the existing thrift types, to keep this diff simpler.

`SharedFeaturizer` now inherits from `Featurizer` and uses the new input types. I created helper functions to make the transition easier, eventually the parse side one will move into some sort of data reader abstraction. For now I simply modified every call site.

I modified the batch implementation to make if it is parallel an implementation detail of an individual featurizer, while explosing a simple batch interface and a default interface that new featurizers can use.

The input structs are at the moment carbon copies of the original thrift struct. This is a transitional feature, and will change. This diff is just to get rid of the implicit api of thrift serialized json and turn it into an explicit api we can actually debate the merits of

Reviewed By: hikushalhere

Differential Revision: D9479075

fbshipit-source-id: a65a69ef9d9ecb6e23d174765892296334b39d47