Expansible Component Config
Summary:
Can add __EXPANSIBLE__ to a Component Class to make it's config in JobSpec automatically be converted to all configs of subclass of the Component (including self).

I didn't enable it for all Component Class because if somebody inherits from a Component which doesn't have any subclass previously, the existing config json file will break, which is error-prone. I thought about make specifying class names in the config json mandatory even if it's not union, but it makes the json very verbose in practice.

Reviewed By: hikushalhere

Differential Revision: D10317560

fbshipit-source-id: f25bc9e6fcbc74669aa3282b7715a96ced680bcd