Allow saving modules in FBLearner
Summary: When running a PyText jobspec on FBLearner with a `save_path` specified for a module, the run with fail with a `PermissionError` when trying to save the module. With this diff I add a `config.modules_save_dir` option, which becomes the base directory in which all modules are saved. When running on FBLearner, this is automatically set to the GFS working directory.

Reviewed By: seayoung1112

Differential Revision: D10383739

fbshipit-source-id: 936a9c5538328a862eebf8a02e0bda79849a66e5