Fine tune the elmo model
Summary:
Only use the last layer of the ELMO inner state for the supervised task. unfreeze the ELMO inner state and use PureDocAttention Model only to reduce the complexity of the model.

The return state of the PureDoc Attention is only the representation, otherwise the `None` states will break the parallel training logic.

=======

Integration test for ELMO fine-tuning `fblearner/flow/run_integration_tests --project assistant.nlu`

P60272620

=======

[V6] Test on CoLA

Boundary First Pool
https://our.intern.facebook.com/intern/fblearner/details/80280925?tab=overview

The finished runs are summerized in the shared quip file https://fb.quip.com/kJsYAxpbmbKK, the best mcc result for CoLA task is 0.52.

==========

[V6] Boundary Pooling is enabled:

combine both bos and eos for the pooling:
https://our.intern.facebook.com/intern/fblearner/details/80280855?tab=overview

Batch size 64, Boundary First Pool,
https://our.intern.facebook.com/intern/fblearner/details/80114990/

Batch size 32, Boundary First Pool,
https://our.intern.facebook.com/intern/fblearner/details/80112558/

Batch size 16, Boundary First Pool,
https://our.intern.facebook.com/intern/fblearner/details/80110500?tab=overview
The per_label_soft_score for `true` case is 0.5214, not as good as the experiment without fine-tuning.
Waiting on the result with both bos and eos for the pooling.

Batch size 8, Boundary First Pool,
https://our.intern.facebook.com/intern/fblearner/details/80106523/

Batch size 4, Boundary First Pool,
https://our.intern.facebook.com/intern/fblearner/details/80106063/

==========

[V4] Use the Boundary Pool instead of the MaxPool

=============

[V2]

We still met CUDA out of memory error afterward. To reduce memory usage: 1) the Attention Model is replace with MaxPool. 2) There is a sweep of batch size: 128, 64, 32, 16, 4 for both train, eval, and test. Except for the batch size of 4, all others still face the memory problem in the training stage.

Here is the flow with MaxPool and batch size of 4.
https://our.intern.facebook.com/intern/fblearner/details/79513277?tab=overview

Here is another run by Haoran for comparison:
https://our.intern.facebook.com/intern/fblearner/details/78780322?tab=Outputs

Result comparison shows that the reduce model has degraded performance for the recall (0.72->0.50)and F1 score (0.76->0.49). Changing BiLSTM to PureDocAttention model to reduce the complexity could be one contributing factor. As the comparison https://our.intern.facebook.com/intern/diff/D10442915/ shows that there is an average of 10% degradation. Another contributing factor is that the removal of the Attention model in the PureDocAttention due to the CUDA out of memory error.

Reviewed By: liaimi

Differential Revision: D10478203

fbshipit-source-id: 0239d983d5c7d868494bbb3a27b1862bef18acce