Install PyTorch Manually (#28)
Summary:
This changes the install instructions to have the user install pytorch manually. The advantage is that it allows the user to pick the correct version of pytorch.
The circle ci tests have been configured to use the cpu version of pytorch.

To test:
I removed pytorch/venv
I followed the install procedures in the updated README.md
I successfully ran the tests.
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/28

Reviewed By: Titousensei, seayoung1112, chenyangyu1988

Differential Revision: D13360256

Pulled By: m3rlin45

fbshipit-source-id: 8c113e42eb69c40c252f7189aca501f368864aa9