Simplify run_tests (#46)
Summary:
This greatly simplifies both the actual command, but also greatly increases the signal when running the tests.
The tests are all still run, but run in a single process which means the results are summarized in a single place. Further, by using the buffer option (`-b`) we suppress most of the output besides those of failing tests.

This diff also lays the groundwork for introducing xml reporting in circleci in the future, by consolidating results into a single process.
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/46

Reviewed By: bethebunny

Differential Revision: D13389301

Pulled By: m3rlin45

fbshipit-source-id: ad7f4a5317a35e4146e4fb26d8eb3fc12f1dc626