Dense feature support, along with related float vector field and decoder input support (#45)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/45

Pretty sure we will eventually decouple models, data_handler and the features we elicit from the configs, so keeping this change simple.

The crux of the change is:
1. New field that can take an array of floats and converts it into a tensor
2. The "input" going into the model's forward is now arranged in the format:
embedding tensors : other input : decoder input
(This is not scalable or maintainable, I know)
3. data_handlers now has to separately collect embedding feature inputs, then finally add the decoder input, if present.

Too bad we can't separate these into true categories via a dict.

Reviewed By: bethebunny

Differential Revision: D13285873

fbshipit-source-id: 468b7a4b73db9847e4c22b6d3caebffcedea65f0