Build actions field with all data by default and don't rebuild vocab. (#58)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/58

Currently, actions field is built with training data only and action label is built with all data. This doesn't pose a problem in functionality because they share the same `ActionsField` object. However, the logging is misleading as it shows that they are not same when eval/test data has extra actions.

Fix this logging error by setting `vocab_from_all_data=True` by default for ActionsField and not rebuilding lael fields if they are built already.

Reviewed By: wowitsmrinal

Differential Revision: D13394031

fbshipit-source-id: f3021055c1f259f5a43c4094eb4d8f6bf9d88c1f