Fixing Python 3.7 incompatibility because of typing module. (#74)
Summary:
Using type() and issubclass() has differences in behaviors between Python 3.6 and 3.7 owing to changes in the typing module. type(Union[X, Y]) no longer returns Union for example, something we were relying on for comparisons. Using __origin__ though fixes it, and this is compatible for 3.6 too.
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/74

Reviewed By: bethebunny

Differential Revision: D13412359

Pulled By: snisarg

fbshipit-source-id: 1e4c318175402af9eb1a4991993377459a004778