Ensure MLP layers on word embeddings are not frozen (#76)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/76

One intended use of MLP layers on word embeddings is to freeze the pretrained embeddings and learn the MLP layers during training. Currently, when the embedding's "freeze" param is True, `create_module()` freezes all of the embedding's param. This diff adds a call to a new `Module.freeze(config)` method, and the `WordEmbedding.freeze` overrides this to only freeze the embedding lookup.

Reviewed By: seayoung1112

Differential Revision: D13386266

fbshipit-source-id: f364b3635dee9ffbb68826c782452aa50792dbb1