gen_default_config command
Summary:
This diff introduces a command to generate a full config with default
values. Users can start with the generated config as a template to get
started, or to discover useful parameters they can tweak.

Introduces a unit test to try and gen all the Task compononents, except the
ones that don't have a default value (to be fixed in later diffs)

Differential Revision: D13248917

fbshipit-source-id: a983e391d2190b119d49e9e1df5339637e2ed6f3