gen_default_config command (#23)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/23

This diff introduces a command to generate a full config with default
values. Users can start with the generated config as a template to get
started, or to discover useful parameters they can tweak.

Introduces a unit test to try and gen all the Task compononents, except the
ones that don't have a default value (to be fixed in later diffs)

Reviewed By: m3rlin45

Differential Revision: D13248917

fbshipit-source-id: 620754e8a128cf7e362b4b4a3c277367b3a4e4e5