Config document generation for RTD (#101)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/101

Pull Request resolved: https://github.com/facebookresearch/pytext/pull/80

Generates rsts for Config objects and adds a Config reference section to the ReadTheDocs site.

The associated component is inferred if possible, and a link is generated to the code documentation for that component (this link process isn't working super well, I think there are issues with classes showing up multiple times in our RTD hierarchy eg. when we transclude classes in __init__.py)

Configs can be marked up as follows:

  class MyConfig(ConfigBase):
    """Normal docstring, usually not relevant..."""

    #: Note the colon at the beginning of this comment
    #: that identifies this as a docstring block to sphinx
    something: int = 4

    #: The number of passes the trainer will make
    #: through the training dataset
    epochs: int = 10

    #: Some other config we need:
    other_config: Other.Config = Other.Config(x=10)

Reviewed By: m3rlin45

Differential Revision: D13393569

fbshipit-source-id: 3b0f6f4de7e3285b1d345bf53e99ea2b95b4312a