Update vocab handling (#133)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/133

Currently the vocab for WordEmbedding comes from the train / eval / test data or a special vocab file. Add a config flag to add all tokens from the pretrained embeddings to the vocab. This can be useful if the training data is small, and we want to keep a larger vocab size for the final model

Also change the parse mode for the pretrained embeddings file to "backslashreplace", to avoid any parsing errors. (anything that would have errored previously gets mapped to a reasonable string, and these are probably very rare tokens anyway).

Reviewed By: bethebunny

Differential Revision: D13472879

fbshipit-source-id: 31c498fef8f4fc0370a016caf33f9d4b52bd67c7