Updating model execution tutorial with DocNN configs
Since this follows immediately after the train tutorial using DocNN, many readers are confused about the outputs seen current on this page as they're expecting to use DocNN, not ATIS that's shown here. We're updating this tutorial to reflect DocNN instead. Although poor in results, the assumption is that someone running ATIS would be better equipped at this stage to understand the results of the test step than a beginner just trying out train and test with PyText.