Using existing config file for export and predict.
We observed a lot of questions regarding the location of the "config.json" file for export and train steps. We expect this file to be created by the user, but for the sake of simplicity and initial trails, it's better to point users to us the existing file. This will also ease users trying the framework for the first time and support questions on both GitHub Issues and FB group.