Reduce duplicate indices (#165)
Summary:
our config objects are being added to the index twice, once in the modules docs, once in the config docs. This change removes the config doc verions from the index. Long term we likely want to merge these, but for now, this will do.

based on discussion here:
https://github.com/sphinx-doc/sphinx/issues/3866
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/165

Reviewed By: bethebunny

Differential Revision: D13553706

Pulled By: m3rlin45

fbshipit-source-id: 62a23cce983e3e3a10a8169c990b9af9c12278bd