Adding dropout for slot-specific values (#155)
Summary:
Sometimes, there may be overfitting on slot values. For example, in [IN:CREATE_CALL call [SL:CONTACT Mom]], the model often overfits on 'Mom' because 'Call Mom' is more common in the training data than calling other contacts.

This change applies a dropout layer to a token embedding (before concatenation with the dictionary features) if the last non-shift, non-reduction oracle (gold) action was a slot (and not an intent).
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/155

Differential Revision: D10843725

Pulled By: emilysw

fbshipit-source-id: 489778682bcad84752e1b19e42227dae6d98496a