Implement attention in StackLSTM for RNNG (Python) (#175)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/175

D10249584 for stack attention was removed during the RNNG pytext file migration, so this diff re-introduces attention for StackLSTM for RNNG (Python).

Attention-weighted prior states are concatenated with the StackLSTM's last state as input into the action classifier. This feature can be turned on or off as an AblationParams flag: use_stack_attention.

Differential Revision: D13558819

fbshipit-source-id: 2c6dd56c6ede5b493388ee31c396aede991d7e0c