Gets rid of all doc gen warnings for Python 3.7 (#188)
Summary:
This fixes the error that was preventing doc generation in Python 3.7. Note that doc generation will still throw a warning for 3.6, but a lot of time has been spent trying to fix that by Chris and me and moving ahead is the best approach anyway. Therefore, this commit also updates doc generation on CircleCI and ReadTheDocs to Python 3.7. Finally, enabling failures on warning for Sphinx doc gen so we lock this down for good.

Tested by building docs locally without errors

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Docs change / refactoring / dependency upgrade
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/188

Reviewed By: hikushalhere

Differential Revision: D13582311

Pulled By: snisarg

fbshipit-source-id: e415d0740f9131ac875ab45183dceab90babb9df