add temperature in kd doc classification
Summary:
reference implementation from https://arxiv.org/pdf/1503.02531.pdf
I'm actually not sure if we need to re-train the teacher using the same temperature. Quoting "In the simplest form of distillation, knowledge is transferred to the distilled model by training it on a transfer set and using a soft target distribution for each case in the transfer set that is produced by using the cumbersome model with a high temperature in its softmax. The same high temperature is used when training the distilled model, but after it has been trained it uses a temperature of 1."

Differential Revision: D13605113

fbshipit-source-id: ba39ff26f165259e17fd3d156bd9485c788dd1f1