add temperature in kd doc classification (#199)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/199

reference implementation from https://arxiv.org/pdf/1503.02531.pdf
I'm actually not sure if we need to re-train the teacher using the same temperature. Quoting "In the simplest form of distillation, knowledge is transferred to the distilled model by training it on a transfer set and using a soft target distribution for each case in the transfer set that is produced by using the cumbersome model with a high temperature in its softmax. The same high temperature is used when training the distilled model, but after it has been trained it uses a temperature of 1."

Reviewed By: gardenia22

Differential Revision: D13605113

fbshipit-source-id: 04dfb2c857db4d41f1b039df4ecfcb9399926683