remove epoch_size to simplify configuration (#207)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/207

The epoch_size parameter to define epochs is counterinutitive from user perspective.  One needs to know dataset size and divide by batch size to judge what the epoch_size should be.  This diff removes this parameter in favor of a binary flag, and a reasonable default epoch size.

Two parameters are added:
- upsample (in DisjointMultitaskDatahandler)
- target_task_name (moved to DisjointMultitask.Config)

If upsample = True (default):
We'll cycle over each dataset repeatedly in round-robin fashion, so that shorter datasets will see more iterations (upsampled).  Epoch is defined to be the epoch of the target task, if defined, otherwise it's set to the length of the shortest dataset.

If upsample = False:
We do a single pass over each dataset.  Datasets which are short will sit idle some of the time.  This is used for evaluation.

Reviewed By: seayoung1112

Differential Revision: D13678696

fbshipit-source-id: 8fa10c2f20da4ce6b7e40f78294c549d2d317a53