default config for JointTextTask
Summary:
because fblearner is confused about types List[Union], we can't
give a default value to those Config parameters and we leave them empty.
That's a problem for gen_default_config, which relies on instanciating
components with default values. So those configs, like JointTextTask, can't
be used with gen_default_config.

This diff introduces a generic solution for Tasks like this (not for all
components, coming soon) by allowing each task to provide an optional
function "example_config" that returns a Config object with default values
suitable for gen_default_config. This default values for List[Union] should
be a list of each possible class of the union, to give a complete example
that users can edit at their convenience.

We solve the case of JointTextTask as an example. Other tasks will be solved
in following diffs.

Differential Revision: D13441705

fbshipit-source-id: cb91d938b84cf78a9bca5f7cb31f73632c2a7f19