Refactor metric_reporters (#245)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/245

we're passing around SummaryWriter everywhere.  This is a
TensorBoard-specific internal parameter that should be isolated to the
TensorBoardChannel class.  This is because metric reporters channels are
actually resources that need to be closed after using them.

This implementation makes it difficult to add new metric reporter channels
that are additions to and alternatives of TensorBoard. This diff refactors
how metric reporters are passed around and improves modularity and
separation of concerns.

Reviewed By: geof90

Differential Revision: D13829608

fbshipit-source-id: e16a25e174510be5aeae1cb38c34460796f4c46b