default config for JointTextTask (#211)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/211

because fblearner is confused about types List[Union], we can't
give a default value to those Config parameters and we leave them empty.
That's a problem for gen_default_config, which relies on instanciating
components with default values. So those configs, like JointTextTask, can't
be used with gen_default_config.

This diff introduces a generic solution for Tasks like this (not for all
components, coming soon) by allowing each task to provide an optional
function "example_config" that returns a Config object with default values
suitable for gen_default_config. This default values for List[Union] should
be a list of each possible class of the union, to give a complete example
that users can edit at their convenience.

We solve the case of JointTextTask as an example. Other tasks will be solved
in following diffs.

Reviewed By: m3rlin45

Differential Revision: D13441705

fbshipit-source-id: 5fe6bb23379a9505351de86954014edf7468733c