Temp fix for beam search top k eval metrics
Summary:
This diff gets rid of the memory surge that occurs when trying to store top k nodes during metrics calculation. It removes the logic that involves processing the trees into nodes for top k.

With this diff, top k frame accuracy will appear on the UI as 0. Meanwhile, we will work on investigating the reason why the top k nodes cause the memory surge.

Differential Revision: D13996416

fbshipit-source-id: eca34742e66aafeb191aebdc67f37d849e8214f0