Temp fix for beam search top k eval metrics (#273)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/273

This diff gets rid of the memory surge that occurs when trying to store top k nodes during metrics calculation. It removes the logic that involves processing the trees into nodes for top k.

With this diff, top k frame accuracy will appear on the UI as 0. Meanwhile, we will work on investigating the reason why the top k nodes cause the memory surge.

Differential Revision: D13996416

fbshipit-source-id: 479be4d90e0d89354ed3595fc5d4303afafaa7ad