Fix pytest unit test failure
Summary:
https://github.com/pytest-dev/pytest/issues/3151

If there isn't a __init__.py file in a directory, pytest imports a test file into the global namespace. This means in a package structure not including __init__.py files, test files in separate directories with the same basename (filename not including directory) will overlap, and fortunately pytest fails with a visible error message rather than not executing the tests.

Differential Revision: D14031515

fbshipit-source-id: a6032e44eda2b32138235e2728925cf56fc24897