Data Handler Rewrite [4/x]: Model functions (#278)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/278

Changes to the Model class that need to be made to support taking inputs from the tensorizers.

I made this on a new Model class so that it is clear how this would be used to implement models which don't conform to the current model structure, but they can probably go on the Model base class instead.

The main important things are arrange_model_inputs, arrange_targets, and train_batch.

Prediction and export functionality will be implemented later.

Reviewed By: borguz

Differential Revision: D13870591

fbshipit-source-id: e081349a97bfeb2e89cc83e2aba06d8b9d7d4cfd