Add flag for training time limit in pytext
Summary:
Add flag for training time limit. If set, then at the beginning of each training epoch, estimate if the next epoch will take the total training time over the time limit. If so, then stop training.

For fblearner jobs, automatically set this limit to just under 3 days. This is helpful since the system would kill the job anyways past that point, and it's convenient to then get the test set evaluation.

Differential Revision: D14138642

fbshipit-source-id: 20c6db24749be19277665291de4593b174903fd2