Bring back topk metrics for RNNG
Summary:
The oom issue was due to the line:

  all_pred_trees: List[List[Tree]] = [[]] * batchSize

which generates n copies of same lists. So in the deleted code in D13996416, it will take O(n2) memory when iterating and invoking tree_to_metric_node.

This was also the reason why fa and topk fa is not the same for k=1

Differential Revision: D14129718

fbshipit-source-id: 3fb46ffe47667dc8aaeaaeb2484371823fcbbc68