Bring back topk metrics for RNNG (#309)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/309

The oom issue was due to the line:

  all_pred_trees: List[List[Tree]] = [[]] * batchSize

which generates n copies of same lists. So in the deleted code in D13996416, it will take O(n2) memory when iterating and invoking tree_to_metric_node.

This was also the reason why fa and topk fa is not the same for k=1

Reviewed By: wowitsmrinal

Differential Revision: D14129718

fbshipit-source-id: 414cdb711d0882b009ba96ebb28887770d1fcc25