Add flag for training time limit in pytext (#314)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/314

Add flag for training time limit. If set, then at the beginning of each training epoch, estimate if the next epoch will take the total training time over the time limit. If so, then stop training.

For fblearner jobs, automatically set this limit to just under 3 days. This is helpful since the system would kill the job anyways past that point, and it's convenient to then get the test set evaluation.

Reviewed By: m3rlin45

Differential Revision: D14138642

fbshipit-source-id: b4de2bf18f1bc3ac73be0075b781254073540be1