implement AugmentedLSTM (#266)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/266

Implements AugmentedLSTM with support for stacking, bidirectionality and variational dropout.

This commit is heavily based off of AllenNLP's implementation of AugmentedLSTM. AugmentedLSTM attempts to be a drop-in replacement for LSTM.

There is one fundamental difference with AllenNLP's implementation which is this implementation fuses all operations in the AugmentedLSTMCell where AllenNLP keeps the operations separate (I assume for readability). Although there is a sacrifice of readability this implementation is roughly 33% faster than AllenNLP's implementation on CPU.

Differential Revision: D13890628

fbshipit-source-id: dc4f39430b10a274f9291d589452ea625aaf02eb