Allow Metric Reporter To Select Metric From Config
Summary: Currently, the metric reporter's metric selection criterior is bounded with the class. It's hard to change it without changing code, and it's hard to apply different one in run time. This diff adds the "select_metric_path" to the metric reporter config. It can be used to select different metric that would like to be used for model selection.

Differential Revision: D14209239

fbshipit-source-id: 2d959e7065f9696853c0b2385a37dab5a92414aa