Add option of hard target to `KLDivergenceCELoss` (#315)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/315

This loss originally tries to match the model's output distribution with a target distribution. This is probably only used for knowledge distillation at the moment, and it can be helpful to also include the true target / label in the loss (according to the originl paper https://arxiv.org/abs/1503.02531). This diff adds a flag `hard_weight` that specifies the ratio of the weights for the hard target to the teacher's distribution.

Reviewed By: jingfeidu

Differential Revision: D14138044

fbshipit-source-id: 478ae67416371a1bdda73614477a08ef29977637