Fix potential bugs in intent-slot metrics (1st)
Summary:
Thanks to Jinfeng who discovered this. `Node` class in intent slot metrics suffers from a potential bug that if its `children` are modified after creation, its hash value may not change accordingly. In this diff we modify the `__hash__()` method of the `Node` class so that it does not rely on `children`. We also add a unit test that would have exposed the bug.

Note current PyText and NLU code paths are not affected by this bug.

Differential Revision: D14196789

fbshipit-source-id: cd8f5ceed2ac05b303a6b22909d50d36d2ef31ef