move batch & pad out of tensorizer (#316)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/316

Data processing needs to do the following:
- Read from a data stream and parse to rows of typed fields
- Process rows (e.g. tokenize, vocab lookup) into numberized rows that can be input to a model
- Batch rows
- Pad a batches of inputs of varying length

Previously batching was done on raw rows out of data source, before tokenization/numberization.  This diff moves batching and padding to take place after tokenization/numberization for the following reason:
- There are batching strategies which depend on the length of tokenized input, for instance packing a batch for language modeling.
- One might want to sort a batch based on tokenized string length, which is also easier this way

The potential drawback of this approach is that it's not clear where to do the padding.  If done in the batcher, the batcher would need to know how to pad and tensorize it's fields, which is not ideal.  For this reason, we pass the batcher output back to tensorizers to pad and tensorize.

Reviewed By: bethebunny

Differential Revision: D14151590

fbshipit-source-id: 3bc3564ad5aac44dde9a6241bbf3401021743858