Fix potential bugs in intent-slot metrics (1st) (#337)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/337

Thanks to Jinfeng who discovered this. `Node` class in intent slot metrics suffers from a potential bug that if its `children` are modified after creation, its hash value may not change accordingly. In this diff we modify the `__hash__()` method of the `Node` class so that it does not rely on `children`. We also add a unit test that would have exposed the bug.

Note current PyText and NLU code paths are not affected by this bug.

Reviewed By: poojasethi, Jeffyrao

Differential Revision: D14196789

fbshipit-source-id: aa8246771137cdab35b8ad720671acb99778ddd5