Multitask: add output view and option to use per-task model_select_metric (#366)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/366

Multitask setup does not produce a flow output - to make inspecting results easier, make the flow output the set of per-task `model_select_metric`s. This is a good choice because this metric is the most important one for each task (e.g. the default for classification is accuracy), and because it's hard to show all metrics for each subtask (which can be of any type of task).

Also, enable the multitask metric reporter to select best epoch according to the sum of `model_select_metric`s as one way of determining best overall epoch.

Reviewed By: borguz

Differential Revision: D14326176

fbshipit-source-id: 9af4bdece7bcc826222117b30d0663cd9e5f18f6