rename RootDataSource functions
Summary:
This diff clarifies the API of RootDataSource and renames the
functions that need to be implemented by RootDataSource users, to make it
clear that it should return a new generator each time.  Previous name can be
understood as the functions iterate on the source data set. Documentation
added or changed accordingly.

Differential Revision: D14508877

fbshipit-source-id: b916d8ace00ad7495727aa70241df65ab7f047fe