rename RootDataSource functions (#405)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/405

This diff clarifies the API of RootDataSource and renames the
functions that need to be implemented by RootDataSource users, to make it
clear that it should return a new generator each time.  Previous name can be
understood as the functions iterate on the source data set. Documentation
added or changed accordingly.

Reviewed By: borguz

Differential Revision: D14508877

fbshipit-source-id: e1c884e7429f11f5af91a32722a3750688f3b84b