Multitask in new data design (#496)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/496

Migrate multitask to the new data design. Specifically, this involves creating a `NewDisjointMultitask` and `NewDisjointMultitaskModel`, and making sure the task context is set in each batch for the model and metric reporter.

Also update `ProbabilisticBatchSampler` to persist data iterators across epochs, so data reading won't restart from the beginning of a file in a new epoch.

Have not added exporting or multi-gpu training for this task yet.

Differential Revision: D14992525

fbshipit-source-id: ce85e9f7f9420522ed1bbb83b920c9ca146505e5