Add Multilingual TSV Data Source (#501)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/501

Currently when we have models which require a language embeddings (for example: finetuning a cross-lingual LM on XNLI), we need to explictly provide the language as a new column to the data source. This is both inefficient (every line is augmented with the language information) as well as inconvenient (we need to create copies of the original datasets with this language information added). In this diff, I'm adding a Multilingual TSV Data source which adds language information to each example.

Reviewed By: borguz

Differential Revision: D15039543

fbshipit-source-id: a7d324db8e351b662a6a44bbdc60e84ecca5c19b