Multitask in new data design (#496)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/496

Migrate multitask to the new data design. Specifically, this involves creating a `NewDisjointMultitask` and `NewDisjointMultitaskModel`, and making sure the task context is set in each batch for the model and metric reporter.

Also update `RandomizedBatchSampler` to persist data iterators across epochs, so data reading won't restart from the beginning of a file in a new epoch.

Have not added exporting or replicated multi-gpu training results for this task yet.

Reviewed By: borguz

Differential Revision: D14992525

fbshipit-source-id: 6327fe45bd18bde3a5a558f1107ad1549c7b342a