TorchScript export for document classification (#503)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/503

Full torchscript export for doc classification.

For now we are not going the route of including the vocab lookup, etc. as part of the model at train time. There's some convenience aspects here, for instance not having to worry about getting logits out of the model at train time and predictions at inference time, but mainly it's a consequence of List/str not being supported by torch tracing. Since we can't trace them and we need them as the inputs to the primary forward function for the model, we're locked in to ScriptModules for every single module in the model graph, which isn't feasible (we don't even control all of the implementations, even if it was already possible to implement everything we want in Script, which it definitely isn't).

As such, we train the model as normal, trace it via jit.trace, and then inject it into a ScriptModule wrapper which implements the vocab lookup and predictions. This module is then saved and exported.

This ends up looking a lot like ONNX/caffe2 wrapping, but thankfully less painful.

Reviewed By: seayoung1112

Differential Revision: D14966825

fbshipit-source-id: 437b4f7f992790df8590c10807513e24f6d9deed