Implement attention for StackLSTM
Summary:
Attention over the stack was accidentally removed during the RNNG pytext file migration; this diff reintroduces attention for StackLSTM for the Python RNNG.

For the stack summary, an attention-weighted embedding of prior StackLSTM states is concatenated with the StackLSTM's last state, to use as input into the ActionLSTM's classifier. Attention can be turned on or off in the config as an AblationParams flag: use_stack_attention.

Differential Revision: D15111094

fbshipit-source-id: dba287232cfb5f033fa9119b73f93d3f92afa60d