Sharded data sources (#456)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/456

Moves sharding from Data to DataSouce, so that we can have DataSources that can shard in different ways.

- Move shard() utility to data/utils
- Introduce ShardedDataSource (base class), RowShardedDataSource, and BlockShardedTSVDataSource

Default behavior is to shard the configured DataSource in a RowShardedDataSource, which results in previous behavior.  RowShardedDataSource is a light wrapper which returns shard(data_source.train)

If the configured DataSource is an instance of ShardedDataSource, then it has sharding capability and we don't wrap it.

BlockShardedTSVDataSource is a TSV data source, which divides the file into blocks and returns csv readers over each block for each shard. This is useful for when row order is important in the file, e.g. when rows are sentences of a long document.  Used for MaskedLM training. Blocks are the same byte size, but not guaranteed to have the same number of rows.  Therefore the Data object needs to set epoch_size for distributed training to work correctly.

Introduce epoch_size for base Data object as an option.  If set, epochs are set to be a fixed number of batches.  Underlying dataset is cycled over infinitely.  If not set, behavior is as before.

Remove epoch size from batch samplers, since it's no longer needed.

Consolidate DisjointMultitaskData and CrossLingualLMData to share more code.

Reviewed By: kartikayk

Differential Revision: D14840136

fbshipit-source-id: e8cad36711a2bef90823c4b0094da17917811493