Enable `CrossLingualData` (renamed) to work with more general data (`MonolingualData`) (#567)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/567

Currently `CrossLingualLMData` only works with `PackedLMData`, but it'd be nice to use this for fine-tuning (via regular `Data` classes). Rename this class to `CrossLingualData` and allow it to sample from `MonolingualData`. This is just a dummy subclass of `Data` so that the config system won't complain about having recursive configs (i.e. `CrossLingualData` having child `Data` objects in the config).

Differential Revision: D15232451

fbshipit-source-id: 9a5f94378bdd872340136c580c299359ed49fb88