Pass raw data to metric_reporter for batch_context (#603)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/603

We have a bunch of pass-through tensorizers that exist to funnel data into the metric reporter. Instead now we pass the raw batch data to metric_reporter.batch_context so it can get that info directly.

This allows us to also delete a bunch of the pass-through tensorizers.

I'll also work to eliminate MetricTensorizer and NtokenTensorizer in a follow-up diff. This is a bit more complicated partly because there are already 10+ models which are using it, partly because it's implemented in the base MetricReporter class, and partly because the number of tokens are located in a different place for different models.

Reviewed By: Titousensei

Differential Revision: D15368631

fbshipit-source-id: d009d8edfe403b824aa916afac09bf826e134651