Migrate `FairseqCharacterTokenizer` to (open-source) `ByteTokenTensorizer`. (#598)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/598

Migrate `FairseqCharacterTokenizer` to open-source `ByteTokenTensorizer`. This tensorizer tokenizes input text and converts each token into a sequence of bytes. The `BiTransformer` model, which was previously the only one to use `FairseqCharacterTensorizer`, will set the field `offset_for_non_padding` to be 1 since it was pre-trained with 1 added to each non-padding byte. This will enable a doc model (in the second diff of the stack) to use concatenated byte
 and word features.

Byte features ideally would be the standard over character features going forward:

- the current `CharacterTokenTensorizer` outputs unicode code points, which can take on over 1 million values. This would require us to have 1 million entries in the character embedding or else it would break.
- byte encoding doesn't need vocab creation, so it's simpler than a fix to `CharacterTokenTensorizer` that creates a vocab of unicode code points
- the performance of the new tensorizer matches the old (see test plan)
- this would share implementation with the bi-transformer's byte tensorizer.

Reviewed By: Titousensei

Differential Revision: D15387086

fbshipit-source-id: dea35a15cb23e4d8432d5e1331b03e1f28c1c164