Prefetch data from Data batch generator on separate thread (#631)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/631

Creating input data tensors for training can be done on CPU asynchronously while forward/backward passes happen on the GPU.

This diff adds a prefetch mechanism for generators which executes the generator asynchronously in a separate thread and waits on them. This can apply generally to any generator, but it's not generally useful due to the python GIL except in cases where the CPU is occasionally waiting on other resources, eg. a database, file, or in PyText's frequent case GPU computations.

We apply this prefetch mechanism to the data batcher, allowing it to generate batches in a separate thread and pre-cache them while the forward/backward GPU passes execute.

In the case where we're training on CPU this should degrade gracefully, as it's only adding the overhead of a thread and threadsafe queue to the data generation, which shouldn't be noticable.

Differential Revision: D15523247

fbshipit-source-id: 121b36df2a0934d84d1e3647f87db2fa69d1e3be