Add config flag to indicate whether model is being loaded from a saved model state (#636)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/636

We need some way to indicate during initialization that the model is being loaded from a saved snapshot model state, so that we can do things such as skipping the loading of pretrained embeddings (since the embedding state is already stored in the saved state).

In the old design, where we still had metadata, we had a somewhat hacky solution: First, the pretrained embeddings were not saved in the metadata in the snapshot (https://fburl.com/skhq8gb8), second, the data handler would skip loading the pretrained embeddings if metadata is already loaded (https://fburl.com/q46dezei, `init_metadata` method is what loads the pretrained embeddings from file).

For the new design, I can think of a few options:

1. Add an extra argument to model from_config function. However, this would require changing every single model in PyText that overrides that function :(

2. Create a global context map that we can use to store something like this. This is probably the better solution, however that will need some design discussion

3. The solution in this diff. While overriding config value in this way is probably a misuse of config, it is a temporary hack that will help me to unblock the migration of Assistant models to the new module design.

Differential Revision: D15534210

fbshipit-source-id: d404589ec18026d221aa0669e576c069f7889730