EASY [PyText] gen-default-config fix for representations (#632)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/632

some classes (like BiLSTMDocAttention, I did not track others)
subclass Module but not Component. This breaks gen-default-config because
it matches the wrong slots.

The reason for that, is looking at matching superclasses, we have to stop in
the hierarchy before the core base classes, otherwise we will match
everything.

This diff fixes it by stopping at Module or Component by default, which
should cover all the PyText classes.

Reviewed By: hikushalhere

Differential Revision: D15524461

fbshipit-source-id: 4f956a6bd67b5873118c8bfcae66b965202b7ea1