Plumb row index to metric reporter (#673)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/673

The row indices were previously plumbed to the metric reporters, whereas in the new behaviour (after D15593875), the indicies are just the indices within the batch. These indices are important in joining with domain info when calculating and evaluating per-domain metrics in the NLU integration tests, which should hopefully be fixed in this diff.

Reviewed By: seayoung1112

Differential Revision: D15722005

fbshipit-source-id: d9c12ec4f8e9339a424996cd5f066ef898367ac6