XLM torchscript model export (#638)
Summary:
Pull Request resolved: https://github.com/facebookresearch/pytext/pull/638

Create the torchscript export for XLM.

We have a couple of different possible APIs.

The core, idealized API sends in per-token language identification. We currently export this from multilingual Yoda, although the accuracy of the per-token languages is questionable (ie. I'm not sure that it actually computes them, I think it just uses the full-sentence language for each token), but we want to support this now in case we can update Yoda to be better in the future.

Then there's a simplified production API for use cases where we're not using multilingual yoda, in which case you can send batched inputs with one language per input in the batch.

There's a final case which I mainly added for ease of integration with benchmark tools, where no language is passed, in which case the language defaults to english for all tokens. It's plausible that this could be useful for real users too, as the accuracy isn't much worse if the language is identified incorrectly.

Reviewed By: kartikayk

Differential Revision: D15504434

fbshipit-source-id: 9bc0e3868bb0fa5c95fb38976041706999692837